Add in-app guide
- Fix a bug in the parser
- Fix the bader_workgraph
